Move build URL to internal extension
